fix(billing): address code review feedback for stripe module
- Hardened Stripe SDK type assertions in webhook handlers
- Soft-deleted instead of hard-deleted pricing plans
- Overhauled and responsive SubmitManagerPricing component
- Centralized billing route exports via an index barrel
- Pinned Stripe SDK to v20.4.1 for type safety
- Adjusted CTAs to route 'Enterprise' plan to contact-sales instead of login
- Implemented payload serializers for Prisma BigInt support
- Added staleTime (5m) for React Query billing hooks

+++ src/features/billing/schemas/billing-schema.ts
@@ -9,7 +9,7 @@ export type CheckoutInput = z.infer<typeof checkoutSchema>
 // ── Pricing Plan CRUD ────────────────────────────────────
 export const pricingPlanCreateSchema = z.object({
   name: z.string().min(1).max(100),
-  price: z.coerce.number().nonnegative(),
+  price: z.coerce.number().int().nonnegative(),
   features: z.record(z.string(), z.boolean()).optional(),
   stripePriceId: z.string().max(255).optional(),
   isActive: z.boolean().default(true),
